pyproject.toml
[tool.poetry]
name = "transformers torch"
version = "0.1.0"
description = "transformers stable zephyr"
authors = ["Alvaro"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.37.2"
torch = "^2.2.0"
tiktoken = "^0.6.0"
accelerate = "^0.27.2"
pyopenssl = "^24.0.0"
huggingface-hub = "^0.20.3"
llama-cpp-python = "^0.2.44"
python-dateutil = "^2.8.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
